pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "unillm"
version = "0.1.0"
description = "A unified interface for using various LLMs"
readme = "README.md"
authors = [
    {name = "Adam Stein", email = "steinad@seas.upenn.edu"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "torch",
    "transformers",
    "openai",
    "anthropic",
    "boto3",
    "python-dotenv",
    "google-genai",
    "pillow"
]

[project.urls]
Homepage = "https://github.com/BrachioLab/universal_llm"